Fixes vulnerability in setStartIndex
Fixes the vulnerability in `setStartIndex` in the contract that allows
for changing the `startIndex` if its value is computed to `0` since this
case is indistinguishable from the value `startIndex` not yet assigned.
Due to restriction of the contract only having access to the last 256
blocks (and the fallback handling for that in `setStartIndex`) the value
of `startIndex` could be modified later if calculated as `0`. Fixes this
issue by initializing `startIndex` to `MAX_TOKENS` (= 0 mod
`MAX_TOKENS`) to distinguish calculated zero from unassigned zero.